Rename module split.py to cv_folds.py
